Added delete button
Added feature where a delete button is added to every row, except for the first row, which deletes the corresponding row. Row numbers are also updated in a separate function. The code that takes in user input has also been updated and put in a separate function to accurately update the user input if a row is deleted